Add pre-refresh capability to caching + docs
Pre-refresh allows for returning the existing cached value, but calling the backing function to refresh the cache proactively to prevent any cache stampedes once a cache line expires. This also includes a tunable alpha option to control the probability of a pre-refresh happening.

Added comprehensive documentation to the project as well.